// src/data/companies.d.ts
export declare const companies: {
    id: number;
    name: string;
    category: string;
    commodity: string[];
    officeAddress: string;
    contactPerson: string;
    phoneNumber: string;
    designation: string;
    emailWebsite: string;
    coordinates: number[];
}[];
export declare const companyCategories: string[];